+++ packages/popmotion-pose/src/factories/pose-setter.ts
@@ -3,37 +3,27 @@ import chain from 'popmotion/compositors/chain';
 import delayAction from 'popmotion/compositors/delay';
 import { flipPose, isFlipPose } from '../dom/flip';
 import { just } from '../inc/default-transitions';
-import { Pose, Poser, PoseSetter, PoseSetterFactoryProps, ChildPoses, Bounds2D, PoseSetterProps } from '../types';
-import { getPoseValues } from 'inc/selectors';
+import {
+  Pose,
+  Poser,
+  PoseSetter,
+  PoseSetterFactoryProps,
+  ChildPoses,
+  Bounds2D,
+  PoseSetterProps
+} from '../types';
+import { getPoseValues } from '../inc/selectors';
 
 export type PoseSetterFactory = (props: PoseSetterFactoryProps) => PoseSetter;
 
-type EnforceBoundary = (v: number) => number;
 type AnimationsPromiseList = Array<Promise<any>>;
 
-type BoundaryMap = {
-  [key: string]: [string, string]
-};
-
-const boundaryMap: BoundaryMap = {
-  x: ['left', 'right'],
-  y: ['top', 'bottom'],
-  z: ['far', 'near']
-};
-
-const addBoundaries = (action: Action, bounds: Bounds2D, key: string): Action => {
-  const enforceBounds: EnforceBoundary[] = [];
-  const [min, max] = boundaryMap[key];
-
-  if (bounds[min] !== undefined) enforceBounds.push((v: number) => Math.max(v, bounds[min]));
-  if (bounds[max] !== undefined) enforceBounds.push((v: number) => Math.min(v, bounds[max]));
-
-  return enforceBounds.length ? action.pipe(...enforceBounds) : action;
-};
-
-const dragPoses = new Set(['dragging', 'dragEnd']);
-
-const startChildAnimations = (children: ChildPoses, next: string, nextPose: Pose, props: PoseSetterProps): Array<Promise<any>> => {
+const startChildAnimations = (
+  children: ChildPoses,
+  next: string,
+  nextPose: Pose,
+  props: PoseSetterProps
+): Array<Promise<any>> => {
   const animations: Array<Promise<any>> = [];
   let delay = 0;
   let stagger = 0;
@@ -46,80 +36,96 @@ const startChildAnimations = (children: ChildPoses, next: string, nextPose: Pose
   }
 
   const maxStaggerDuration = children.size - 1 * stagger;
-  const generateStaggerDuration = staggerDirection === 1
-    ? (i: number) => i * stagger
-    : (i: number) => maxStaggerDuration - (i * stagger);
+  const generateStaggerDuration =
+    staggerDirection === 1
+      ? (i: number) => i * stagger
+      : (i: number) => maxStaggerDuration - i * stagger;
 
   Array.from(children).forEach((child: Poser, i: number) => {
-    animations.push(child.set(next, {
-      ...props,
-      delay: delay + generateStaggerDuration(i)
-    }));
+    animations.push(
+      child.set(next, {
+        ...props,
+        delay: delay + generateStaggerDuration(i)
+      })
+    );
   });
 
   return animations;
 };
 
-const createPoseSetter: PoseSetterFactory = (setterProps) => (next, props = {}) => {
+const createPoseSetter: PoseSetterFactory = setterProps => (
+  next,
+  props = {}
+) => {
   const { delay = 0 } = props;
-  const { activeActions, activePoses, children, poses, values, types, dragProps, flipEnabled } = setterProps;
+  const {
+    activeActions,
+    activePoses,
+    children,
+    poses,
+    values,
+    types,
+    dragProps,
+    flipEnabled
+  } = setterProps;
   const animations: AnimationsPromiseList = [];
   const { dragBounds } = dragProps;
   let nextPose = poses[next];
 
   if (nextPose) {
-    if (flipEnabled && isFlipPose(nextPose, next)) nextPose = flipPose(setterProps, nextPose);
+    if (flipEnabled && isFlipPose(nextPose, next))
+      nextPose = flipPose(setterProps, nextPose);
     const { transition: getTransition } = nextPose;
 
     const poserAnimations = Object.keys(getPoseValues(nextPose)).map(
-      (key) => new Promise((complete) => {
-        const value = values.get(key);
-        const type = types.get(key);
-        const from = value.get();
-        const unparsedTarget = nextPose[key];
+      key =>
+        new Promise(complete => {
+          const value = values.get(key);
+          const type = types.get(key);
+          const from = value.get();
+          const unparsedTarget = nextPose[key];
 
-        // Stop the current action
-        if (activeActions.has(key)) activeActions.get(key).stop();
+          // Stop the current action
+          if (activeActions.has(key)) activeActions.get(key).stop();
 
-        // Get transition if `transition` prop isn't false
-        let transition = getTransition({
-          ...props,
-          from: type ? type.parse(from) : from,
-          velocity: value.getVelocity() || 0,
-          to: type ? type.parse(unparsedTarget) : unparsedTarget,
-          key,
-          prevPoseKey: activePoses.get(key)
-        });
+          // Get transition if `transition` prop isn't false
+          let transition = getTransition({
+            from: type ? type.parse(from) : from,
+            velocity: value.getVelocity() || 0,
+            to: type ? type.parse(unparsedTarget) : unparsedTarget,
+            key,
+            prevPoseKey: activePoses.get(key),
+            dragBounds,
+            ...props
+          });
 
-        if (transition === false) transition = just(unparsedTarget);
+          if (transition === false) transition = just(unparsedTarget);
 
-        // If this is a drag pose, apply boundaries
-        if (dragPoses.has(next) && dragBounds && boundaryMap[key]) transition = addBoundaries(transition, dragBounds, key);
+          // Add delay if defined
+          if (delay || nextPose.delay) {
+            transition = chain(
+              delayAction(delay || nextPose.delay),
+              transition
+            );
+          }
 
-        // Add delay if defined
-        if (delay || nextPose.delay) {
-          transition = chain(
-            delayAction(delay || nextPose.delay),
-            transition
-          );
-        }
+          // Start transition
+          const transitionApi = transition.start({
+            update: (v: any) => value.update(v),
+            complete
+          });
 
-        // Start transition
-        const transitionApi = transition.start({
-          update: (v: any) => value.update(v),
-          complete
-        });
-
-        activeActions.set(key, transitionApi);
-        activePoses.set(key, next);
-      })
+          activeActions.set(key, transitionApi);
+          activePoses.set(key, next);
+        })
     );
 
     animations.push(...poserAnimations);
   }
 
   // Get children animation Promises
-  if (children.size) animations.push(...startChildAnimations(children, next, nextPose, props));
+  if (children.size)
+    animations.push(...startChildAnimations(children, next, nextPose, props));
 
   return Promise.all(animations);
 };