+++ packages/popmotion-pose/src/inc/default-transitions.ts
@@ -3,19 +3,41 @@ import tween from 'popmotion/animations/tween';
 import action, { Action } from 'popmotion/action';
 import pointer from 'popmotion/input/pointer';
 import { eachValue } from './transition-composers';
-import { RawValue, Transition } from '../types';
+import { RawValue, Transition, Bounds2D } from '../types';
 
-const singleAxisPointer = (axis: string) => (from: number) => pointer({ [axis]: from }).pipe((v: any) => v[axis]);
+const singleAxisPointer = (axis: string) => (from: number) =>
+  pointer({ [axis]: from }).pipe((v: any) => v[axis]);
 const pointerX = singleAxisPointer('x');
 const pointerY = singleAxisPointer('y');
 
-export const just = (from: RawValue): Action => action(({ update, complete }) => {
-  update(from);
-  complete();
-});
+const createPointer = (
+  axisPointerCreator: (from: number) => Action,
+  min: string,
+  max: string
+): Transition => ({ from, dragBounds }) => {
+  const axisPointer = axisPointerCreator(from);
+  const transformQueue: Array<(v: number) => number> = [];
 
-const underDampedSpring: Transition = ({ from, velocity, to }) => spring({ from, to, velocity, stiffness: 500, damping: 25 });
-const overDampedSpring: Transition = ({ from, velocity, to }) => spring({ from, to, velocity, stiffness: 700, damping: 35 });
+  if (dragBounds) {
+    if (dragBounds[min] !== undefined)
+      transformQueue.push((v: number) => Math.max(v, dragBounds[min]));
+    if (dragBounds[max] !== undefined)
+      transformQueue.push((v: number) => Math.min(v, dragBounds[max]));
+  }
+
+  return transformQueue.length ? axisPointer.pipe(...transformQueue) : action;
+};
+
+export const just = (from: RawValue): Action =>
+  action(({ update, complete }) => {
+    update(from);
+    complete();
+  });
+
+const underDampedSpring: Transition = ({ from, velocity, to }) =>
+  spring({ from, to, velocity, stiffness: 500, damping: 25 });
+const overDampedSpring: Transition = ({ from, velocity, to }) =>
+  spring({ from, to, velocity, stiffness: 700, damping: 35 });
 
 // TODO: Adjust transitions based on behavioural props
 const intelligentTransition: Transition = eachValue({
@@ -28,10 +50,9 @@ const intelligentTransition: Transition = eachValue({
   default: tween
 });
 
-// TODO: Move add boundaries here
 const dragAction: Transition = eachValue({
-  x: ({ from }) => pointerX(from as number),
-  y: ({ from }) => pointerY(from as number)
+  x: createPointer(pointerX, 'left', 'right'),
+  y: createPointer(pointerY, 'top', 'bottom')
 });
 
 // TODO: Return `decay` based on behavioural props