+++ packages/react-pose/src/components/PoseGroup.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export class PoseGroup extends React.PureComponent {}