+++ packages/react-pose/src/global.js
@@ -1,3 +1,0 @@
-import { Pose } from './';
-
-window.Pose = Pose;