+++ packages/react-pose/src/global.ts
@@ -1,0 +1,4 @@
+import posed, { PoseGroup } from './index';
+
+(window as any).posed = posed;
+(window as any).PoseGroup = PoseGroup;