+++ packages/react-pose/src/index.js
@@ -1,3 +1,0 @@
-import Pose from './Pose';
-
-export { Pose };