+++ packages/react-pose/src/index.ts
@@ -1,0 +1,5 @@
+import posed from './posed';
+import PoseGroup from './components/PoseGroup';
+
+export default posed;
+export { PoseGroup };