+++ site/styles/themes.js
@@ -1,24 +1,24 @@
-import menus from '~/data/menus.json';
-import content from '~/data/content.json';
-import PopmotionLogo from '~/components/icons/Logo';
-import PoseLogo from '~/components/icons/PoseLogo';
-import { color } from './vars';
+import menus from "~/data/menus.json";
+import content from "~/data/content.json";
+import PopmotionLogo from "~/components/icons/Logo";
+import PoseLogo from "~/components/icons/PoseLogo";
+import { color } from "./vars";
 
 const generateTheme = (name, props) => ({
   data: {
     menus: menus[name],
     content: content[name],
-    rootUrl: name === 'popmotion' ? '/' : `/${name}`
+    rootUrl: name === "popmotion" ? "/" : `/${name}`
   },
   ...props
 });
 
 const themeSettings = {
   popmotion: {
-    id: 'popmotion',
-    name: 'Popmotion',
-    tagline: 'A functional JavaScript motion library',
-    sections: ['api', 'learn', 'blog', 'community'],
+    id: "popmotion",
+    name: "Popmotion",
+    tagline: "A functional JavaScript motion library",
+    sections: ["api", "learn", "blog", "community"],
     Logo: PopmotionLogo,
     homepageLogoSize: {
       width: 283,
@@ -31,10 +31,10 @@ const themeSettings = {
     mastheadBackground: `linear-gradient(${color.brand}, ${color.brandBurn})`
   },
   pose: {
-    id: 'pose',
-    name: 'Popmotion Pose',
-    tagline: 'A declarative motion library for HTML, SVG and React',
-    sections: ['api', 'learn', 'community'],
+    id: "pose",
+    name: "Popmotion Pose",
+    tagline: "A declarative motion system for HTML, SVG and React",
+    sections: ["api", "learn", "community"],
     Logo: PoseLogo,
     homepageLogoSize: {
       width: 253,